+++ packages/core/integration-tests/test/html.js
@@ -940,6 +940,10 @@ describe('html', function () {
         assets: ['300x300.png'],
       },
     ]);
+
+    let contents = await outputFS.readFile(b.getBundles()[0].filePath, 'utf8');
+    assert(contents.includes('src="/'));
+    assert(contents.includes('srcset="/'));
   });
 
   it('should detect srcset attribute of source element', async function () {
@@ -968,7 +972,7 @@ describe('html', function () {
 
     const html = await outputFS.readFile(b.getBundles()[0].filePath, 'utf8');
 
-    const source = html.match(/<source srcset=".*>/)[0];
+    const source = html.match(/<source srcset="\/.*>/)[0];
 
     assert(source.split(', ').length === 3);
   });
@@ -996,6 +1000,9 @@ describe('html', function () {
         assets: ['300x300.png'],
       },
     ]);
+
+    let contents = await outputFS.readFile(b.getBundles()[0].filePath, 'utf8');
+    assert(contents.includes('imagesrcset="/'));
   });
 
   it.skip('should support webmanifest', async function () {