+++ packages/transformers/js/src/JSTransformer.js
@@ -101,6 +101,7 @@ type MacroAsset = {|
 // NOTE: Make sure this is in sync with the TypeScript definition in the @parcel/macros package.
 type MacroContext = {|
   addAsset(asset: MacroAsset): void,
+  loc: SourceLocation,
   invalidateOnFileChange(FilePath): void,
   invalidateOnFileCreate(FileCreateInvalidation): void,
   invalidateOnEnvChange(string): void,
@@ -194,9 +195,7 @@ export default (new Transformer({
             });
         }
 
-        if (automaticJSXRuntime) {
-          jsxImportSource = reactLib;
-        }
+        jsxImportSource = reactLib;
       }
 
       isJSX = Boolean(compilerOptions?.jsx || pragma);
@@ -340,7 +339,9 @@ export default (new Transformer({
       type,
       jsx_pragma: config?.pragma,
       jsx_pragma_frag: config?.pragmaFrag,
-      automatic_jsx_runtime: Boolean(config?.automaticJSXRuntime),
+      automatic_jsx_runtime: Boolean(
+        config?.automaticJSXRuntime || asset.meta.jsxRuntime === 'automatic',
+      ),
       jsx_import_source: config?.jsxImportSource,
       is_development: options.mode === 'development',
       react_refresh: Boolean(config?.reactRefresh),
@@ -426,6 +427,19 @@ export default (new Transformer({
                       specifier: k,
                       specifierType: 'esm',
                     });
+                  },
+                  loc: {
+                    filePath: asset.filePath,
+                    start: {
+                      line:
+                        loc.start_line + Number(asset.meta.startLine ?? 1) - 1,
+                      column: loc.start_col,
+                    },
+                    end: {
+                      line:
+                        loc.end_line + Number(asset.meta.startLine ?? 1) - 1,
+                      column: loc.end_col,
+                    },
                   },
                   invalidateOnFileChange(filePath) {
                     asset.invalidateOnFileChange(filePath);